discover ./web-ext-config.js (#1)
