+++ example/web-ext-config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  run: {
+    startUrl: ["https://www.google.com"],
+  }
+};
